move load_schema to PMOChecker; added load_schema_by_version; added a global __schema__ so default only has to be changed in 1 place;
